First attempt at "rendering" images. DOES NOT WORK!
The idea was using the "\r" character to delete and reprint the lines. This appears
to work on short lines–less that the terminal window width–but I have been unable
to make this work with longer lines (although the patch responsible for compiling
a string appropriately does appear to work).